tts-service/requirements.txt
piper-tts==1.4.1
fastapi==0.115.12
uvicorn[standard]==0.34.0
numpy==2.2.3
pydantic==2.10.6
